Remove 'controller' module
