+++ bee-network/src/service/mod.rs
@@ -4,7 +4,6 @@
 #![cfg(feature = "full")]
 
 pub mod command;
-pub mod controller;
 pub mod error;
 pub mod event;
 pub mod service;
